Added the start of an idea I had whereby a single sketch could be added as a child of a HText object.  The path along the sketch could influence the placement of the characters.  The framework is there and it certainly influences the placement of characters but the effect could not be described as 'good' (yet).
I have commented out the entry point (GetTools() method) so
that it doesn't get faults logged against it until it produces
something more pleasant.

The way I have implemented it is to add an OrientationModifier
class.  This is a layer between HText and the sketch which
adds the functionality of converting a 'distance' from the
HText::glCommands() method into a gp_Trsf transformation
matrix that can be applied to the coordinates used to
render the text.  I half wondered whether this
OrientationModifier class might be used to allow placement of
other sketch objects along a base sketch's path.  Maybe an
'array along' sort of thing.  That's all just glossy brochure stuff at the moment.