Added drop-down menu items to the right mouse button while in drawing mode.  These menu options allow the user to simulate left mouse button clicks at exact locations using existing graphics.  eg: it you have a square sketch already and you're adding a circle, instead of left clicking where the circle's centre should go, you can right click on the sketch and select the 'Click midpoint' option.  This will be just like clicking the left mouse button at the midpoint of the square.
If you right mouse click over the end of a line or on a point, the point menu has an extra 'Offset from point' option.  When this option is selected, a separate dialog box opens asking for an offset from the chosen point's location.  eg: if you want to draw something 50 to the right of an existing point (or end of line etc.) then right click on the point, select the 'offset from point' menu option and then press the OK button.  This will be like clicking the left mouse button exactly 50 to the right of the selected point.

Finally, I have added the same 'offset from last point' option to the Input properties dialog.  This allows the operator to give coordinates relative to the last 'clicked' position.  Again, this only exists in one of the drawing modes.

There are also 'Click on' options for the top, bottom, left, right and middle of a circle as well as centre of an arc.